RC5: added support for extended RC5
Extended RC5 uses the second start bit as the inverse of the most
significant bit of the command to extended the command length to
seven bits.

When sending, if the number of bits is 12 or fewer, we send a second
start bit (1).  If the number of bits is larger than 12, then we invert
the most significant bit and don't send the second start bit.

When receiving, we only expect one start bit and then invert the next
received bit.  This gives the correct result for both RC5 and extended
RC5, although the number of received bits will be 13 in both cases.